Add session middleware to the API
With the session middleware in place, sessions are started automatically
if necessary and all requests handled by the API now send session
cookies in the response.

The session middleware can be configured with the new options for the
filename of the database to use and for some some coookie settings. The
implementation of this is a bit awkward, though, because the setup
function in intelmq_manager/serve.py cannot add new middleware, so we
have to modify an already added middleware in place.

Another problem with the current implementation is that the path for the
cookie is not set, so the browser will end up with many cookies, one for
each path used in the requests to the API.