Use one sqlite3 connection per API server
Instead of opening new sqlite connections for each request we now open
only one per server and use a lock to serialize access to from multiple
threads.

To avoid blocking connections to the same database from other processes,
we now open the connection in auto-commit mode as a simple way to ensure
that no transactions are open for very long. For our purposes
auto-commit it fine because so far the transactions consist of single
SQL statements anyway. The only exception is initialization but there we
now explicitly start a transaction.

Also, with auto-commit there's not much point for now to distinguis
between queries and transactions in the convenience methods so we
consolidate the query and transaction methods into one execute method.