Implement token based authentication
The server now requires authentication: The client has login with the
new login API function and then use the returned session token in
subsequent requests.

This commit implements the API part of this, using new reworked session
store for the user accounts and token implementation. The client is
modified to have a login button that uses a modal dialog to ask for
username and password, fetches a token from the server and stores it in
the browser's sessionStorage. The JS-initiated HTTP requests are now
done through the new authenticatedGetJson function in static.js which
retrieves the token and sends it in the Authorization header.

The implementation is still a little rough, particularly the UI part. In
particular, authentication is effectively optional in the sense that if
no session store has been configured, no login is required. However the
login button is always shown. Also the styling of login button and
dialog could be improved and some of the behavior could be better (e.g.
log out should return to the main page).