Implement session expiration
Whenever a session token is looked up, the session store now first
deletes all sessions that have expired. Also, when the session token is
verified its timestamp in the session store is updated to reset the
expiration timer.

The duration of the sessions can be configured with the new option
session_duration.